Add test pipeline (#11)
* Add test pipeline

* Add build step in test pipeline

* Improve GitVersion integration for build number retrieval

Refactor the GetBuildNumber task to enhance the handling of GitVersion. The script now checks if GitVersion.Tool is installed and retrieves the build number accordingly. If GitVersion is unavailable or fails, it defaults to "1.0.0". This change ensures more reliable versioning in the build process.

* Fix module path filtering in showcase step

Updated the module path filtering logic to exclude source files when importing the built module. This change ensures that only the relevant module is loaded, improving the reliability of the showcase step in the test pipeline. Added a log message to display all .psd1 files found if the module is not located.

* Remove duplicate files

* Refactor variable and path management in pipeline

Updated the variable and path management sections in the test pipeline.
Improved logging for environment variables and added verification steps to ensure variables and paths are set correctly.

* Update tools path in test pipeline

Changed the tools path from 'tools/bin' to 'build' in the test pipeline configuration.

* Refactor showcase commands in test pipeline

Updated the showcase commands in the test pipeline to improve clarity and organization. Sections were simplified, and comments were added for better understanding. Enhanced output messages for better visibility during execution. This refactor aims to streamline the showcase process and improve maintainability.

* Improve logging format in PowerShell setup

Updated the logging output for PowerShell version and variable verification steps to enhance readability. The changes include adding section markers for better visibility in the logs.

* Simplify pipeline type handling in Write-PipelineSection

Removed the switch statement for pipeline type handling, since both GitHub Actions and Azure Pipelines support ##[section]

* Remove command log type from Write-PipelineLog

Eliminated the 'command' log type from the Write-PipelineLog function to ensure it behaves the same way in both environments

* Refactor command showcase in test pipeline

Updated the command showcase section in the test pipeline to improve clarity and organization. Each command is now properly numbered and categorized, enhancing readability and understanding of the showcased functionalities.

* Fix command log type from Write-PipelineLog

Fixed the 'command' log type in the Write-PipelineLog function to ensure it behaves the same way in both environments

* Fix output formatting in Write-PipelineSection

Removed trailing newline from the output message in the Write-PipelineSection function.

* Update command message in Write-PipelineCommand

Replaced the command message in Write-PipelineCommand to provide a more appropriate example of command execution

* Add empty line after section header output

This change adds an empty line after writing the section header to the pipeline log. This improves readability by visually separating log entries, making it easier for users to follow the output in CI/CD pipeline runs.

* Update task completion status in test pipeline

Changed the task completion message to indicate partial success
by updating the status to 'SucceededWithIssues'. This provides
more accurate feedback during the pipeline execution.

* Improve handling of Secret and ReadOnly flags in GitHub Actions

Updated the Set-PipelineVariable function to replace Write-PipelineLog with Write-PipelineWarning for better clarity when the Secret and ReadOnly flags are used in GitHub Actions. The Secret flag now masks the value in logs and sets it appropriately.

* Add Azure DevOps pipeline and showcase script

This commit introduces a new Azure DevOps pipeline configuration file and a PowerShell script to showcase all commands from the PipelineUtils module. The pipeline is designed to trigger on pull requests and main branch updates, ensuring that the commands are tested in both GitHub Actions and Azure Pipelines environments.

* Fix typo

* Enhance Set-PipelineReleaseNumber for Build pipelines

Added a check to ensure Set-PipelineReleaseNumber is only used in Azure DevOps Release pipelines. If invoked in a Build pipeline, it now issues a warning and sets the build number instead. Updated tests to verify this behavior.

* Refactor tests to improve output handling

Updated test assertions in Set-PipelineVariable, Write-PipelineLog, and Write-PipelineSection to join output arrays into strings for better comparison.

* Fix output handling

* Refactor Write-PipelineSection for GitHub Actions

Added a conditional check to ensure an empty line is printed
for GitHub Actions pipelines. This improves log readability
by maintaining consistent output formatting across different
pipeline types.

+semver: major